# Conflicts:
#	Leaf.java

+++ Leaf.java
@@ -6,8 +6,10 @@
 public class Leaf extends NodeEntry
 {
     private int dataBlockID; //Block index that points to a specific block in dataFile.
-    private long recordID; //Record index that points into a Record in a specific block.
-    Leaf(int dataBlockID,long recordID, MBR mbr) {
+    private int recordID; //Record index that points into a Record in a specific block.
+
+
+    Leaf(int dataBlockID,int recordID, MBR mbr) {
         super(mbr);
         this.dataBlockID = dataBlockID;
         this.recordID = recordID;
@@ -26,7 +28,7 @@ public class Leaf extends NodeEntry
      * Getter function that return the record index in the dataBlock saved in dataFile.
      * @return integer index of a record in current block stored in DataFile.
      */
-    long getRecordID()
+    int getRecordID()
     {
         return recordID;
     }
